initial support for jobs
some general simplifications (e.g. removed 'undeployChannel')

added support for stream module parameters

added support for stream module classloader

named channel support in stream definitions

handling job undeployment

When a stream is created (but not deployed) a znode will be created in the following format:

/xd/streams/stream-name

When the stream is deployed, /source, /sink/, and
(optionally) /processors znodes will be created, with
their children being the labels for the modules of those
types that belong to the stream.

When the stream is undeployed, the admin will delete
the /source, /sink, and /processors nodes to bring
the znode back to its original undeployed state:

/xd/streams/stream-name